Fix for the tests
